Fix 9 pre-release audit issues (HIGH + MEDIUM severity)
FIX 1 (HIGH): Serialization.cpp — stop writing nativeSampleRate to .bsp;
it was persisted but never deserialized, polluting every saved project.

FIX 2 (HIGH): ExportRenderer.cpp — promote numSamples from int to int64_t
to prevent silent truncation on songs longer than ~12 minutes at 48 kHz.

FIX 3 (MEDIUM): ArrangementResolver.cpp — pre-roll all playChance gates
before computing stackGain so simultaneous-stack normalisation uses the
actual survivor count rather than the pre-filter block count.

FIX 4 (MEDIUM): EntryMixer.h — advance destOff/destCount in the
srcStart < 0 branch to preserve time alignment when lead-in starts
before the window origin.

FIX 5 (MEDIUM): InspectorPanel.cpp — stack-count +/−/remove buttons now
capture a pre-state snapshot and call applyExternalMutation so all edits
are undoable via Ctrl+Z.

FIX 6 (MEDIUM): MainComponent.cpp — playBlock/playClip set timelinePos =
startMark (not 0) so the lead-in region plays correctly and totalDuration
covers endMark, matching the resolver's convention.

FIX 7 (MEDIUM): PlaybackEngine.h — update class doc to accurately describe
CriticalSection + tryEnter pattern (was claiming lock-free atomic swap).

FIX 8 (MEDIUM): InspectorPanel.cpp — remove !b->isDone guard from
maxPlayable count so Done blocks are not incorrectly excluded from the
stack-play-count cap (isDone is cosmetic only).

FIX 9 (MEDIUM): InspectorPanel.cpp/h — delete all dead overlapSlider code:
tooltip setup, setVisible calls, sliderValueChanged branch, drag-start/end
listeners, resized layout block, and the three field declarations.

+++ Source/Audio/ExportRenderer.cpp
@@ -20,18 +20,18 @@ bool ExportRenderer::renderToFile(const ResolvedArrangement& arrangement,
     // FIX H4: guard against int overflow for very long arrangements
     if (totalSamples > (int64_t)std::numeric_limits<int>::max()) return false;
 
-    const int numChannels = 2;
-    const int numSamples  = (int)totalSamples;
+    const int     numChannels = 2;
+    const int64_t numSamples  = totalSamples;  // int64_t; cast to int only at point of use
 
     // Allocate output buffer
-    juce::AudioBuffer<float> output(numChannels, numSamples);
+    juce::AudioBuffer<float> output(numChannels, (int)numSamples);
     output.clear();
 
     // FIX H1: use shared mixEntryToBuffer so export and playback are identical.
     // Export always runs at project sample rate (pToH = hToP = 1.0, currentHead = 0).
     const int numEntries = arrangement.entries.size();
     for (int i = 0; i < numEntries; ++i) {
-        mixEntryToBuffer(arrangement.entries.getReference(i), output, numSamples, 0LL, 1.0, 1.0, i);
+        mixEntryToBuffer(arrangement.entries.getReference(i), output, (int)numSamples, 0LL, 1.0, 1.0, i);
         if (progress) progress((float)(i + 1) / (float)numEntries);
     }
 
@@ -49,8 +49,7 @@ bool ExportRenderer::renderToFile(const ResolvedArrangement& arrangement,
                                0));
     if (!writer) return false;
 
-
-    return writer->writeFromAudioSampleBuffer(output, 0, numSamples);
+    return writer->writeFromAudioSampleBuffer(output, 0, (int)numSamples);
 }
